Add `cross_centered` helper func
