+++ crates/yanu/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod gui;
+pub mod utils;
